Adding a facebook controller, was getting ambigious ref on the facebook get/post
