+++ FacebookTest/bin/FacebookTest.dll.config
@@ -13,6 +13,8 @@
   <system.web>
     <compilation debug="true" targetFramework="4.6.1" />
     <httpRuntime targetFramework="4.6.1" />
+    <customErrors mode="Off">
+    </customErrors>
     <httpModules>
       <add name="ApplicationInsightsWebTracking" type="Microsoft.ApplicationInsights.Web.ApplicationInsightsHttpModule, Microsoft.AI.Web" />
     </httpModules>
